Added many sample data for the database Backend - added api endpoints to interract with database
